Добавил обработчики команд, CommandFactory, добавил сообщения о неправильном типе id и year, переделал тип id на Long

+++ src/main/java/org/example/DAO/BookDAO.java
@@ -45,7 +45,7 @@ public class BookDAO {
     /**
      * Получает книгу по её ID
      */
-    public Book getById(int id) {
+    public Book getById(Long id) {
         try (Session session = factory.getCurrentSession()) {
             session.beginTransaction();
             Book book = session.get(Book.class, id);
@@ -69,7 +69,7 @@ public class BookDAO {
     /**
      * Удаляет книгу из базы данных по ID
      */
-    public void deleteBook(int id) {
+    public void deleteBook(Long id) {
         try (Session session = factory.getCurrentSession()) {
             session.beginTransaction();
             Book book = session.get(Book.class, id);
